pyproject.toml
[tool.poetry]
name = "telegram-parser-bot"
version = "0.1.0"
description = "Telegram channel parser and bot for application processing"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
aiogram = "^3.1.1"
telethon = "^1.34.0"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

